Back rolling statistics package with the watermill crate
Replaces the hand-rolled EwmaStream/RollingStream with adapters over
`watermill` (generic serializable online statistics, a Rust port of
river.stats). Two small nodes in nodes/statistics.rs:

- StatStream<S: Univariate<f64>>       — wraps any self-contained stat
  (EWMean, RollingMin, RollingMax, RollingQuantile).
- WindowStatStream<R: RollableUnivariate<f64>> — owns a VecDeque window
  around a revertable stat (Sum, Mean, Variance), replicating
  watermill's evict-then-update logic (its own Rolling wrapper borrows
  the stat `&mut` and can't be embedded in a long-lived node).

Same fluent API (ewma, rolling_{sum,mean,min,max,var,std}) plus
rolling_median(n), which watermill's sorted-window quantile gives us for
free. Behavioural notes vs the earlier hand-rolled version: rolling
var/std return 0.0 (not NaN) before two samples, and EWMean seeds on the
first sample.

Refs #111
