+++ wingfoil/Cargo.toml
@@ -96,6 +96,7 @@ serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 crossbeam = "0.8.4"
 num-traits = "0.2"
+watermill = "0.1.2"
 derive-new = "0.7"
 serde-aux = "4.7.0"
 tynm = "0.2.0"
